fix(ir): replace #1431 cast fold with construction-time rejection
#1431 tried to fix a hardware bug where pto.tcvt corrupts values when invoked
with target_dtype == src_dtype (the instruction is designed for cross-dtype
conversion only). The original fix folded the no-op cast away in Simplify, but
the fold escaped its scope and erased programmer-written shaped-type aliases
like the mi = mi_next seam in flash-attention pool loops that MemoryReuse
relies on as an SSA boundary, corrupting cross-iteration state on NPU.

Rather than retry the fold with progressively narrower guards, reject the
malformed op at construction time:

- DeduceTensorCastType: CHECK target_type != input dtype.
- DeduceTileCastType:   CHECK target_type != input dtype.

The error names the offending dtype and tells the user to remove the cast or
pick a different target_type. Catching this at construction keeps malformed
casts out of every downstream pass, codegen, and runtime path — the IR design
itself disallows what the hardware can't execute.

Latent same-dtype casts uncovered by the new CHECK:

- examples/models/qwen3_jit/kernels/rmsnorm.py: post_rmsnorm's docstring claims
  resid is BF16, but the qwen3_decode caller passes resid1_tile (FP32). Drop
  the no-op cast(..., FP32) at both loop entries; the math accepts FP32
  directly. Docstring updated to match actual usage.
- tests/ut/jit/test_spmd.py: the test fixture fed FP32 through pl.cast(..., FP32).
  Switch the fixture to BF16 so the cast is a real BF16->FP32 promotion (the
  test itself is about JIT alpha-renaming under SPMD, the cast is incidental).

Reverts the Simplify cast-fold from this branch's prior commit (the fold is no
longer needed once construction rejects the input).

+++ tests/ut/jit/test_spmd.py
@@ -146,7 +146,10 @@ def test_jit_spmd_sibling_pipeline_loops_reuse_names():
         return out
 
     post = entry.compile_for_test(
-        torch.empty([_BATCH, _HIDDEN], dtype=torch.float32).normal_(),
+        # Input is BF16 so the `pl.cast(..., FP32)` in `rmsnorm_like` is a
+        # real cross-dtype promotion. Same-dtype casts are rejected at IR
+        # construction time, so this fixture must not feed FP32 in.
+        torch.empty([_BATCH, _HIDDEN], dtype=torch.bfloat16).normal_(),
         torch.empty([_BATCH, _HIDDEN], dtype=torch.bfloat16),
     )
     func_types = {f.func_type for f in post.functions.values()}
